requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
openai
faiss-cpu
pydantic 